separated tests for eputils
